first primitive archetype for exchange client apis

+++ api/src/main/java/at/outdated/bitcoin/exchange/api/currency/CurrencyValue.java
@@ -108,9 +108,8 @@ public class CurrencyValue {
         return this;
     }
 
-    public CurrencyValue divide(long div) {
-        value /= (double) div;
-
+    public CurrencyValue divide(double div) {
+        value /= div;
         return this;
     }
